[Misc]: Refactor `wallet_core_rs` crate
* Remove `tw_proto` test FFIs
