+++ rust/tw_proto/Cargo.toml
@@ -10,8 +10,6 @@ fuzz = ["arbitrary"]
 # Enable in fuzz tests only!
 arbitrary = { version = "1", features = ["derive"], optional = true }
 quick-protobuf = "0.8.1"
-tw_encoding = { path = "../tw_encoding" }
-tw_memory = { path = "../tw_memory" }
 
 [build-dependencies]
 pb-rs = "0.10.0"
