+++ rust/wallet_core_rs/src/ffi/ethereum/mod.rs
@@ -2,7 +2,5 @@
 //
 // Copyright © 2017 Trust Wallet.
 
-#[cfg(feature = "ethereum-abi")]
 pub mod abi;
-#[cfg(feature = "ethereum-rlp")]
 pub mod rlp;
