+++ rust/wallet_core_rs/src/ffi/mod.rs
@@ -2,8 +2,11 @@
 //
 // Copyright © 2017 Trust Wallet.
 
+#[cfg(feature = "bitcoin")]
 pub mod bitcoin;
+#[cfg(feature = "ethereum")]
 pub mod ethereum;
+#[cfg(feature = "solana")]
 pub mod solana;
-#[cfg(feature = "uuid")]
-pub mod uuid_ffi;
+#[cfg(feature = "utils")]
+pub mod utils;
