+++ rust/wallet_core_rs/src/ffi/utils/mod.rs
@@ -2,5 +2,5 @@
 //
 // Copyright © 2017 Trust Wallet.
 
-pub mod address;
-pub mod transaction;
+pub mod uint;
+pub mod uuid_ffi;
