+++ rust/wallet_core_rs/src/ffi/utils/uint.rs
@@ -1,6 +1,3 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 // Copyright © 2017 Trust Wallet.
-
-pub mod address;
-pub mod transaction;
