+++ rust/wallet_core_rs/src/ffi/utils/uuid_ffi.rs
